# [PERF] account: Fix OOM when opening Secure Entries wizard